// src/app/modules/hcm/shared/models/loans/loan-type.ts
export class LoanType {
    idCompany: number = -1;
    idLoanType: number = -1;
    idTransactionType: number = -1;
    idInitialSetup: boolean = true;
    idUserCreate: number = -1;
    idUserUpdate: number = -1;
    name: string = "";
    createDate: string = "";
    updateDate: string = "";
    active: boolean = true
}